requirements.txt
# Airflow (install separately or via Docker)
# apache-airflow==2.9.3

# ML / Data
scikit-learn>=1.5
pandas>=2.2
numpy>=1.26
xgboost>=2.1

# Experiment tracking
mlflow>=2.15

# Serialization
joblib>=1.4

# Configuration
pyyaml>=6.0
python-dotenv>=1.0

# API serving
flask>=2.2
gunicorn>=22.0

# Testing
pytest>=8.3
pytest-cov>=5.0
pytest-mock>=3.14
